Relations demo: better suggestions to avoid Admin library conflicts
When updating the objectbox package, projects that use the Admin library
typically also need to manually update that library to avoid a runtime
error.
However, when using a version range for the objectbox package running
pub upgrade may silently upgrade it. So suggest using a concrete
version and add a note to pubspec to not forget about updating the
Admin library.

This matches changes to the Admin online documentation.